setting up task suite
