+++ task-manager/tests/fixtures/db.js
@@ -1,0 +1,25 @@
+const mongoose = require('mongoose')
+const jwt = require('jsonwebtoken')
+const User = require('../../src/models/user')
+
+const userOneID = new mongoose.Types.ObjectId()
+const userOne = {
+    _id: userOneID,
+    name: "User One Ieva",
+    email: "ieva.aleksandravica+userone@gmail.com",
+    password: "userOneIeva123",
+    tokens: [{
+        token: jwt.sign({ _id: userOneID }, process.env.JWT_SECRET)
+    }]
+}
+
+const setupDatabase = async () => {
+    await User.deleteMany()
+    await new User(userOne).save()
+}
+
+module.exports = {
+    userOneID,
+    userOne,
+    setupDatabase
+}
